requirements.txt
asgiref==3.9.1
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
Django==5.2.6
django-cors-headers==4.8.0
django-environ==0.12.0
django-filter==25.1
django-stubs==5.2.9
django-stubs-ext==5.2.9
djangorestframework==3.16.1
djangorestframework-stubs==3.16.8
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
packaging==25.0
pillow==12.1.1
psycopg2-binary==2.9.10
PyJWT==1.7.1
pyTelegramBotAPI==4.29.1
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
sqlparse==0.5.3
types-PyYAML==6.0.12.20250915
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
whitenoise==6.11.0
